refactor(web): W10 slice 1 — lib/release/storefront-* misfiling to lib/storefront (BI-AB17E1A8)
Move-only: the five storefront modules (actions/auth/data/middleware/types
+ 4 test files) leave lib/release/ for the 83-file lib/storefront/ home the
pass §3.1-b table calls "pure misfiling". git mv, no behavior change.

Mechanical follow-through in the same slice:
- 7 @/lib/release/storefront-* import/mock specifiers rewritten
  (2 storefront API routes + tests, lib/storefront/public-trust,
  2 components/storefront sections);
- the 5 legacy root shims (lib/storefront-*.ts) now re-export from
  ./storefront/* — shim retirement (33 importers) stays out of scope;
- lib/release barrel + barrel test trimmed to the branding modules that
  remain in the release domain;
- scripts/check-endpoint-classification.mjs PROXY_CLASSIFIER_REL + its
  test follow the file in the SAME commit (its zero-prefix extraction
  fails loudly on a stale path) — self-test 10/10, guard green
  (264 governed, 3 classified, 261 grandfathered — unchanged);
- path-accuracy comment fixes in lib/ea/route-extract.ts and
  lib/shared/org-address.ts.

Characterization: lib/release + lib/storefront + app/api/storefront +
components/storefront vitest 89 files / 741 tests green after the move
(pre-move 14-dir baseline 181 files / 2001 tests green, recorded).
Application-boundary guard unchanged: 8 contexts, 531 cross-context
imports, 63 exceptions (release/storefront are not governed contexts).

Docs-Impact-Decision: move-only refactor; no user-facing surface changed;
the pass doc already names this exact move (§3.1-b, move #7).
Design-Grounding-Decision: grounded in
docs/architecture/2026-08-16-simplify-and-strengthen-architecture-pass.md
§3.1-b + §5 W10 (BI-AB17E1A8); substrate reviewed —
apps/web/lib/storefront/ (destination home, no name collisions),
scripts/check-endpoint-classification.mjs proxy-classifier coupling,
legacy lib/storefront-*.ts shims kept as pointers.
Process-Spine-Decision: executes a documented architecture-pass move
(§3.1-b storefront misfiling); mechanical git mv + import rewrites only.
