+++ apps/web/lib/release/index.ts
@@ -1,9 +1,6 @@
 // lib/release/index.ts — IT4IT S5.5 Release (AGT-ORCH-500)
 // Barrel export for the release domain module.
-export * from "./storefront-actions";
-export * from "./storefront-data";
-export * from "./storefront-types";
-export * from "./storefront-middleware";
-export * from "./storefront-auth";
+// The storefront-* modules moved to lib/storefront/ (Simplify & Strengthen
+// W10 — they were pure misfiling beside the 83-file lib/storefront/ home).
 export * from "./branding";
 export * from "./branding-presets";
