+++ apps/web/lib/storefront-actions.ts
@@ -1,2 +1,2 @@
-// Shim — moved to lib/release/storefront-actions.ts (Phase 6 refactoring)
-export * from "./release/storefront-actions";
+// Shim — moved to lib/storefront/storefront-actions.ts (Phase 6 refactoring)
+export * from "./storefront/storefront-actions";
