+++ apps/web/lib/storefront-data.ts
@@ -1,2 +1,2 @@
-// Shim — moved to lib/release/storefront-data.ts (Phase 6 refactoring)
-export * from "./release/storefront-data";
+// Shim — moved to lib/storefront/storefront-data.ts (Phase 6 refactoring)
+export * from "./storefront/storefront-data";
