+++ apps/web/lib/storefront-middleware.ts
@@ -1,2 +1,2 @@
-// Shim — moved to lib/release/storefront-middleware.ts (Phase 6 refactoring)
-export * from "./release/storefront-middleware";
+// Shim — moved to lib/storefront/storefront-middleware.ts (Phase 6 refactoring)
+export * from "./storefront/storefront-middleware";
